Making the metricsdb files to keep around configurable.
Currently, the number of most recent metricsdb files to keep around is
determined by a code level constant. In some cases there might be a
requirement to keep more files around (or for the reader to never delete
them). This change introduces a config for that in the
performance-analyzer.properties file named 'max-metrics-db-files'.
Setting this to value -1, makes the files to never be deleted,
otherwise, the acceptable numbers are between 1 and 1000. Anything else
sets the config to default which is 4.

Github issue #47

+++ src/main/java/com/amazon/opendistro/elasticsearch/performanceanalyzer/metricsdb/MetricsDB.java
@@ -261,13 +261,19 @@ public class MetricsDB implements Removable {
         conn.commit();
     }
 
+    /**
+     * We delete the physical file on disk, only if the user has not set the DB_FILE_MAX_COUNT_CONF_NAME
+     * to -1. Otherwise we just close the connection and return.
+     **/
     @Override
     public void remove() throws Exception {
        conn.close();
-       File dbFile = new File(getDBFilePath());
-       if (!dbFile.delete()) {
-           LOG.error("Failed to delete File - {} with ExceptionCode: {}", getDBFilePath(), StatExceptionCode.OTHER.toString());
-           StatsCollector.instance().logException();
+       if (PluginSettings.instance().getMaxMetricsDBFilesCount() != PluginSettings.DB_FILE_VALUE_FOR_NO_DELETION) {
+          File dbFile = new File(getDBFilePath());
+          if (!dbFile.delete()) {
+              LOG.error("Failed to delete File - {} with ExceptionCode: {}", getDBFilePath(), StatExceptionCode.OTHER.toString());
+              StatsCollector.instance().logException();
+          }
        }
     }
